refactor: move download logic out to its own component
also move logic out to a service
