--- packages/studio-web/src/app/demo/demo.component.sass
@@ -1,6 +1,0 @@
-.download__buttons
-    display: flex
-    button.download
-        margin-top: 5px
-        margin-left: 15px
-    justify-content: flex-end
