+++ packages/studio-web/src/app/ras.service.ts
@@ -34,11 +34,18 @@ export interface SupportedLanguage {
   names: { [code: string]: string };
 }
 
+export interface ReadAlongSlots {
+  title: string;
+  subtitle: string;
+}
+
 export enum SupportedOutputs {
-  eaf,
-  textgrid,
-  srt,
-  vtt,
+  eaf = "eaf",
+  textgrid = "textgrid",
+  srt = "srt",
+  vtt = "vtt",
+  html = "html",
+  zip = "zip",
 }
 
 @Injectable({
@@ -50,7 +57,7 @@ export class RasService {
 
   convertRasFormat$(
     body: ReadAlongFormatRequest,
-    output_type: SupportedOutputs | string,
+    output_type: SupportedOutputs,
   ): Observable<Blob> {
     return this.http.post(
       this.baseURL + "/convert_alignment/" + output_type,
