+++ packages/studio-web/src/app/shared/download/download.component.html
@@ -1,0 +1,19 @@
+<div class="download__buttons">
+  <mat-form-field class="d-none d-lg-block" appearance="fill">
+    <mat-label i18n="Output format">Output Format</mat-label>
+    <mat-select [(ngModel)]="selectedOutputFormat">
+      <mat-option *ngFor="let format of outputFormats" [value]="format.value">
+        {{ format.display }}
+      </mat-option>
+    </mat-select>
+  </mat-form-field>
+  <button
+    class="download"
+    color="primary"
+    (click)="download()"
+    mat-mini-fab
+    aria-label="Example icon button with a download icon"
+  >
+    <mat-icon>file_download</mat-icon>
+  </button>
+</div>
