+++ packages/studio-web/src/app/shared/download/download.component.sass
@@ -1,0 +1,6 @@
+.download__buttons
+    display: flex
+    button.download
+        margin-top: 5px
+        margin-left: 15px
+    justify-content: flex-end
