+++ packages/studio-web/src/app/shared/download/download.component.spec.ts
@@ -1,0 +1,22 @@
+import { ComponentFixture, TestBed } from "@angular/core/testing";
+
+import { DownloadComponent } from "./download.component";
+
+describe("DownloadComponent", () => {
+  let component: DownloadComponent;
+  let fixture: ComponentFixture<DownloadComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [DownloadComponent],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(DownloadComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it("should create", () => {
+    expect(component).toBeTruthy();
+  });
+});
