+++ packages/studio-web/src/app/shared/download/download.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from "@angular/core/testing";
+
+import { DownloadService } from "./download.service";
+
+describe("DownloadService", () => {
+  let service: DownloadService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(DownloadService);
+  });
+
+  it("should be created", () => {
+    expect(service).toBeTruthy();
+  });
+});
